requirements.txt
asgiref==3.11.0
dj-database-url==3.1.2
Django==6.0
django-csp==4.0
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
